+++ tests/conftest.py
@@ -50,3 +50,44 @@ def assert_time(t, hour, minute, second, microsecond=None):
 
     if microsecond is not None:
         assert microsecond == t.microsecond
+
+
+def assert_duration(dur, years=None, months=None, weeks=None, days=None,
+                    hours=None, minutes=None, seconds=None,
+                    microseconds=None):
+    expected = {}
+    actual = {}
+
+    if years is not None:
+        expected['years'] = dur.years
+        actual['years'] = years
+
+    if months is not None:
+        expected['months'] = dur.months
+        actual['months'] = months
+
+    if weeks is not None:
+        expected['weeks'] = dur.weeks
+        actual['weeks'] = weeks
+
+    if days is not None:
+        expected['days'] = dur.remaining_days
+        actual['days'] = days
+
+    if hours is not None:
+        expected['hours'] = dur.hours
+        actual['hours'] = hours
+
+    if minutes is not None:
+        expected['minutes'] = dur.minutes
+        actual['minutes'] = minutes
+
+    if seconds is not None:
+        expected['seconds'] = dur.remaining_seconds
+        actual['seconds'] = seconds
+
+    if microseconds is not None:
+        expected['microseconds'] = dur.microseconds
+        actual['microseconds'] = microseconds
+
+    assert expected == actual